Major Change:
Previously, tracking changes across multiple threads (i.e. multiple RLMRealm instances) was not supported and would result in unexpected behavior: once one Realm instance committed changes, then all changes logged up until that point (could be from different threads) for that Realm path were passed along.

To fix this problem, there is now a new class called: RBQRealmChangeLogger which creates logger instances that are associated with a specific Realm instance/thread. When the associated Realm commits the changes, then the changes are passed to RBQRealmNotificationManager which passes it to any listeners. The notification manager is now acting as a similar object to NSNotificationManager, and simply registers and unregisters notifications.